refactor: make popular collection components reusable
