+++ resources/js/Pages/Collections/Components/CollectionsFilterPopover/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CollectionsFilterPopover";
