+++ resources/js/Pages/Collections/Components/CollectionsFilterTabs/index.tsx
