+++ resources/js/Pages/Collections/Components/CollectionsSortingTabs/index.tsx
@@ -1,0 +1,1 @@
+export * from "./CollectionsSortingTabs";
