+++ resources/js/Pages/Collections/Components/MyCollectionsHeading/MyCollectionsHeading.tsx
@@ -1,14 +1,17 @@
 import { Trans, useTranslation } from "react-i18next";
+import { RefreshButton } from "./RefreshButton";
 import { Heading } from "@/Components/Heading";
 import { formatFiat } from "@/Utils/Currency";
 import { tp } from "@/Utils/TranslatePlural";
 
-export const PopularCollectionsHeading = ({
+export const MyCollectionsHeading = ({
     stats,
     currency,
+    wallet,
 }: {
     stats: App.Data.Collections.CollectionStatsData;
     currency: string;
+    wallet: App.Data.Wallet.WalletData | null;
 }): JSX.Element => {
     const { t } = useTranslation();
 
@@ -19,13 +22,15 @@ export const PopularCollectionsHeading = ({
                     level={1}
                     className="dark:text-theme-dark-50"
                 >
-                    {t("pages.popular_collections.title")}
+                    {t("common.my_collections")}
                 </Heading>
+
+                <RefreshButton wallet={wallet} />
             </div>
 
-            <span className="mt-2 block text-sm font-medium leading-5 text-theme-secondary-700 dark:text-theme-dark-200 sm:text-base">
+            <span className="mt-1 block text-sm font-medium leading-5 text-theme-secondary-700 dark:text-theme-dark-200 sm:text-base">
                 <Trans
-                    i18nKey="pages.popular_collections.header_title"
+                    i18nKey="pages.collections.header_title"
                     values={{
                         nftsCount: stats.nfts,
                         collectionsCount: stats.collections,
