+++ resources/js/Pages/Collections/Components/MyCollectionsHeading/index.tsx
@@ -1,0 +1,1 @@
+export * from "./MyCollectionsHeading";
