--- resources/js/Pages/Collections/Components/PopularCollectionsFilterPopover/index.tsx
@@ -1,1 +1,0 @@
-export * from "./PopularCollectionsFilterPopover";
