--- resources/js/Pages/Collections/Components/PopularCollectionsHeading/index.tsx
@@ -1,1 +1,0 @@
-export * from "./PopularCollectionsHeading";
