--- resources/js/Pages/Collections/Components/PopularCollectionsSorting/index.tsx
@@ -1,1 +1,0 @@
-export * from "./PopularCollectionsSorting";
