+++ resources/js/Pages/Collections/Hooks/useCollectionFilters.ts
@@ -4,14 +4,14 @@ import { router } from "@inertiajs/react";
 import { useEffect, useState } from "react";
 import { useDebounce } from "@/Hooks/useDebounce";
 import { useIsFirstRender } from "@/Hooks/useIsFirstRender";
-import { type ChainFilter, type PeriodFilterOptions } from "@/Pages/Collections/Components/PopularCollectionsFilters";
-import { type PopularCollectionsSortBy } from "@/Pages/Collections/Components/PopularCollectionsSorting";
+import { type ChainFilter, type PeriodFilterOptions } from "@/Pages/Collections/Components/CollectionsFilterTabs";
+import { type CollectionsSortByOption } from "@/Pages/Collections/Components/CollectionsSortingTabs";
 
 export type SortByDirection = "asc" | "desc";
 
 export interface Filters extends Record<string, FormDataConvertible> {
     chain?: ChainFilter;
-    sort?: PopularCollectionsSortBy;
+    sort?: CollectionsSortByOption;
     period?: PeriodFilterOptions;
     query?: string;
     perPage?: number;
@@ -20,7 +20,7 @@ export interface Filters extends Record<string, FormDataConvertible> {
 
 interface CollectionFiltersState {
     currentFilters: Filters;
-    setSortBy: (sort: PopularCollectionsSortBy | undefined) => void;
+    setSortBy: (sort: CollectionsSortByOption | undefined) => void;
     setChain: (sort: ChainFilter | undefined) => void;
     setPeriod: (period: PeriodFilterOptions | undefined) => void;
     searchQuery: string;
@@ -73,7 +73,7 @@ export const useCollectionFilters = ({
         }));
     };
 
-    const setSortBy = (sort: PopularCollectionsSortBy | undefined, direction?: SortByDirection): void => {
+    const setSortBy = (sort: CollectionsSortByOption | undefined, direction?: SortByDirection): void => {
         setCurrentFilters((filters) => ({
             ...filters,
             period: sort === undefined ? filters.period : undefined,
